# openapi/integration-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Сервис ремонта оборудования — интеграционный API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Интеграционный API, версия 1"
    }
  ],
  "paths": {
...
    "/requests/{request_id}/attachments": {
      "post": {
        "tags": [
          "attachments"
        ],
        "summary": "Upload Attachment",
        "description": "Multipart с полем `file`. Повтор с тем же `Idempotency-Key` и тем же содержимым возвращает прежний результат; другой файл под тем же ключом — `409 IDEMPOTENCY_CONFLICT`.\n\nТребует scope `requests:write`.",
        "operationId": "upload_attachment_requests__request_id__attachments_post",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_attachment_requests__request_id__attachments_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentView"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не пройдена аутентификация.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или недоступен вызывающему.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния: версия, идемпотентность или недопустимый переход.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка проверки полей.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен, повторите позже.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Файл больше допустимого размера.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Формат файла не поддерживается (тип определяется по содержимому).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка. Формат одинаковый для всех методов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-scopes": [
          "requests:write"
        ]
      }
    },
    "/attachments/{attachment_id}/content": {
      "get": {
        "tags": [
          "attachments"
        ],
        "summary": "Download Attachment",
        "description": "Вложение заявки — по scope `requests:read`, вложение карточки биржи — по `marketplace:read`: какой нужен, определяется владельцем файла.\n\nТребует scope `requests:read` или `marketplace:read`.",
        "operationId": "download_attachment_attachments__attachment_id__content_get",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "attachment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attachment Id"
            }
          },
          {
            "name": "variant",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "safe",
              "title": "Variant"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Некорректный запрос.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не пройдена аутентификация.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или недоступен вызывающему.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния: версия, идемпотентность или недопустимый переход.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка проверки полей.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен, повторите позже.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка. Формат одинаковый для всех методов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-scopes": [
          "requests:read",
          "marketplace:read"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttachmentView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_kind": {
            "type": "string",
            "title": "Owner Kind"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Id"
          },
          "equipment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment Id"
          },
          "slot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot"
          },
          "visibility_class": {
            "type": "string",
            "title": "Visibility Class"
          },
          "processing_state": {
            "type": "string",
            "title": "Processing State"
          },
          "publication_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication State"
          },
          "rejected_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejected Reason"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "byte_size": {
            "type": "integer",
            "title": "Byte Size"
          },
          "pixel_width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pixel Width"
          },
          "pixel_height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pixel Height"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_kind",
          "visibility_class",
          "processing_state",
          "mime_type",
          "byte_size",
          "created_at"
        ],
        "title": "AttachmentView"
      },
...
      "Body_upload_attachment_requests__request_id__attachments_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "slot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot"
          },
          "message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Id"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_attachment_requests__request_id__attachments_post"
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id",
              "details"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Машиночитаемый код ошибки, например NOT_FOUND."
              },
              "message": {
                "type": "string",
                "description": "Сообщение на русском для лога/UI."
              },
              "request_id": {
                "type": "string",
                "description": "Идентификатор запроса — тот же, что в заголовке X-Request-ID."
              },
              "details": {
                "type": "object",
                "description": "Дополнительные поля ошибки, зависят от кода.",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Ключ интеграции сервисной компании: `Authorization: Bearer <ключ>`.",
        "scheme": "bearer",
        "bearerFormat": "rk_<env>_<prefix>_<secret>"
      }
    }
  }
}